moved cluster_ips to commons
