--- fle/cluster/__init__.py
@@ -1,3 +1,0 @@
-from .local.cluster_ips import get_local_container_ips
-
-__all__ = ['get_local_container_ips']
