+++ fle/eval/algorithms/beam/run_parallel.py
@@ -11,7 +11,7 @@ from dotenv import load_dotenv
 
 from fle.agents.formatters import RecursiveReportFormatter
 from fle.agents.llm.api_factory import APIFactory
-from fle.cluster import get_local_container_ips
+from fle.commons.cluster_ips import get_local_container_ips
 from fle.commons.db_client import DBClient, create_db_client
 from fle.commons.models.game_state import GameState
 from fle.env import FactorioInstance
