+++ fle/eval/algorithms/mcts/__main__objective_mcts.py
@@ -12,7 +12,7 @@ from rich import print
 from fle.agents.formatters.conversation_formatter_abc import (
     PLANNING_ADDITION_PROMPT, StructurePreservingFormatter)
 from fle.agents.llm.api_factory import APIFactory
-from fle.cluster import get_local_container_ips
+from fle.commons.cluster_ips import get_local_container_ips
 from fle.commons.db_client import DBClient
 from fle.commons.models.conversation import Conversation
 from fle.commons.models.game_state import GameState
