+++ fle/eval/algorithms/mcts/__main__planning_agent.py
@@ -20,7 +20,7 @@ from typing import List, Tuple
 from fle.agents.formatters.conversation_formatter_abc import \
     PLANNING_ADDITION_PROMPT
 from fle.agents.llm.api_factory import APIFactory
-from fle.cluster import get_local_container_ips
+from fle.commons.cluster_ips import get_local_container_ips
 from dotenv import load_dotenv
 from rich import print
