+++ fle/eval/algorithms/mcts/main.py
@@ -1,10 +1,10 @@
 import os
 import asyncio
 from dotenv import load_dotenv
-from fle.cluster import get_local_container_ips
+from fle.commons.cluster_ips import get_local_container_ips
 from fle.commons.db_client import DBClient
-from eval.algorithms.mcts import MCTSFactory
-from eval.open.plots.run_results import RunResults
+from fle.eval.algorithms.mcts import MCTSFactory
+from fle.eval.open.plots.run_results import RunResults
 from fle.env import FactorioInstance
 import concurrent.futures
 from typing import List, Tuple
